Update virtual-machines-size-specs.md
(The complete list of valid options should potentially be validated with the team who owns the sizes - these simply worked for me.)  When allocating the A* sizes in Azure Resource Manager, you need to use Standard_A* or Basic_A* (some only support standard). When trying to allocate with simply "A8" as an example, this did not work.